Split tilesets.* into individual tilesets. There is one for each of the four main general purpose graphics banks, and one for each of the five map tilesets. Addresses are added to addr/addr_en.js. The other addr files have not been updated yet. Only decoding from the normal en locations is supported so far.
The last of the map tilesets is in the same bank as the last of the general purpose tilesets. They have been mostly separated, but a portion of the general purpose tileset has been allowed to duplicate into the map tileset to avoid special casing its lenght.

Changes are mostly in romSplitter.py, and spriteSheet.py. The romSplitter.py functonality might make more sense in tilesets.py and spriteSheet.py is probably too tightly coupled with how the tilesets are stored, but refactoring those seemed too big of a project.